+++ optimizations/JPEG/Program.cs
@@ -13,8 +13,8 @@ namespace JPEG
 	class Program
 	{
 		const int CompressionQuality = 70;
-		private static readonly IReadOnlyList<Func<Pixel,double>> ChannelSelectors
-			= new Func<Pixel, double>[] { p => p.Y, p => p.Cb, p => p.Cr };
+		private static readonly IReadOnlyList<Func<Pixel,float>> ChannelSelectors
+			= new Func<Pixel, float>[] { p => p.Y, p => p.Cb, p => p.Cr };
 
 		static void Main(string[] args)
 		{
@@ -94,9 +94,9 @@ namespace JPEG
 				{
 					for (var x = 0; x < image.Width; x += DCTSize)
 					{
-						var _y = new double[DCTSize, DCTSize];
-						var cb = new double[DCTSize, DCTSize];
-						var cr = new double[DCTSize, DCTSize];
+						var _y = new float[DCTSize, DCTSize];
+						var cb = new float[DCTSize, DCTSize];
+						var cr = new float[DCTSize, DCTSize];
 						foreach (var channel in new []{_y, cb, cr})
 						{
 							var quantizedBytes = new byte[DCTSize * DCTSize];
@@ -115,7 +115,7 @@ namespace JPEG
 			return result;
 		}
 
-		private static void ShiftMatrixValues(double[,] subMatrix, int shiftValue)
+		private static void ShiftMatrixValues(float[,] subMatrix, int shiftValue)
 		{
 			var height = subMatrix.GetLength(0);
 			var width = subMatrix.GetLength(1);
@@ -125,7 +125,7 @@ namespace JPEG
 					subMatrix[y, x] = subMatrix[y, x] + shiftValue;
 		}
 
-		private static void SetPixels(Matrix matrix, double[,] a, double[,] b, double[,] c, PixelFormat format, int yOffset, int xOffset)
+		private static void SetPixels(Matrix matrix, float[,] a, float[,] b, float[,] c, PixelFormat format, int yOffset, int xOffset)
 		{
 			var height = a.GetLength(0);
 			var width = a.GetLength(1);
@@ -135,9 +135,9 @@ namespace JPEG
 					matrix.Pixels[yOffset + y, xOffset + x] = new Pixel(a[y, x], b[y, x], c[y, x], format);
 		}
 
-		private static double[,] GetSubMatrix(Matrix matrix, int yOffset, int yLength, int xOffset, int xLength, Func<Pixel, double> componentSelector)
+		private static float[,] GetSubMatrix(Matrix matrix, int yOffset, int yLength, int xOffset, int xLength, Func<Pixel, float> componentSelector)
 		{
-			var result = new double[yLength, xLength];
+			var result = new float[yLength, xLength];
 			for(var j = 0; j < yLength; j++)
 				for(var i = 0; i < xLength; i++)
 					result[j, i] = componentSelector(matrix.Pixels[yOffset + j, xOffset + i]);
@@ -174,7 +174,7 @@ namespace JPEG
 			};
 		}
 
-		private static byte[,] Quantize(double[,] channelFreqs, int quality)
+		private static byte[,] Quantize(float[,] channelFreqs, int quality)
 		{
 			var result = new byte[channelFreqs.GetLength(0), channelFreqs.GetLength(1)];
 
@@ -190,9 +190,9 @@ namespace JPEG
 			return result;
 		}
 
-		private static double[,] DeQuantize(byte[,] quantizedBytes, int quality)
+		private static float[,] DeQuantize(byte[,] quantizedBytes, int quality)
 		{
-			var result = new double[quantizedBytes.GetLength(0), quantizedBytes.GetLength(1)];
+			var result = new float[quantizedBytes.GetLength(0), quantizedBytes.GetLength(1)];
 			var quantizationMatrix = GetQuantizationMatrix(quality);
 
 			for(int y = 0; y < quantizedBytes.GetLength(0); y++)
